server/requirements.txt
anyio==3.7.1
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==41.0.1
ecdsa==0.18.0
fastapi==0.99.1
fastapi-sessions==0.3.2
greenlet==3.0.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
packaging==23.1
passlib==1.7.4
pluggy==1.2.0
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.11
pytest==7.4.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.22.0
